requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.20.9
psycopg2==2.8.6
python-dateutil==2.8.1
SQLAlchemy==1.3.11
Werkzeug==0.16.0
